+++ .github/workflows/dev.yaml
@@ -19,7 +19,7 @@ jobs:
           python-version: '3.7'
           architecture: 'x64'
       - name: 'Setup environment'
-        run: make venv
+        run: make env-dev
       - name: 'Run format check'
         run: make format-check
       - name: 'Run unit tests'
