--- requirements-dev.txt
@@ -1,5 +1,0 @@
-black
-pytest
-pytest-cov
-moto
-pylint
