+++ requirements.txt
@@ -1,3 +1,10 @@
 s3fs==0.4.2
 pandas
 pyarrow
+
+# dev
+black
+pytest
+pytest-cov
+moto
+pylint
